Particleクラスの修正とWinMain関数の初期化値変更
- Particle::Update()でindexをfloatにキャストするよう変更。
- Particle.hから不要な#include <thread>を削除。
- WinMain関数でpGameEyeの初期化値を更新。

+++ Engine/Features/Particle/Particle.cpp
@@ -50,7 +50,7 @@ void Particle::Update()
     {
         transforms_[index].scale = Vector3(-1.0f, 1.0f, 1.0f);
         transforms_[index].rotate = Vector3(0.0f, 0.0f, 0.0f);
-        transforms_[index].translate = Vector3(0.0f, 0.0f, index);
+        transforms_[index].translate = Vector3(0.0f, 0.0f, static_cast<float>(index));
         Matrix4x4 wMatrix = Matrix4x4::AffineMatrix(transforms_[index].scale, transforms_[index].rotate, transforms_[index].translate);
         instancingData_[index].world = wMatrix;
         instancingData_[index].wvp = wMatrix * pGameEye_->GetViewProjectionMatrix();
